fix the same for bit-merge
